Display contract interaction/deployment properly
