+++ app/src/containers/WalletHistory/WalletHistory.vue
@@ -86,8 +86,10 @@ import {
   ACTIVITY_STATUS_PENDING,
   ACTIVITY_STATUS_SUCCESSFUL,
   ACTIVITY_STATUS_UNSUCCESSFUL,
+  CONTRACT_INTERACTION_KEY,
   CONTRACT_TYPE_ERC20,
   CONTRACT_TYPE_ERC721,
+  DEPLOY_CONTRACT_ACTION_KEY,
   MAINNET,
   TOKEN_METHOD_APPROVE,
 } from '../../utils/enums'
@@ -189,7 +191,12 @@ export default {
       }
     },
     getActionText(activity) {
-      // Handling tx from common-api schema and /tx schema separately.
+      if (activity.transaction_category === CONTRACT_INTERACTION_KEY) {
+        return this.t('walletActivity.contractInteraction')
+      }
+      if (activity.transaction_category === DEPLOY_CONTRACT_ACTION_KEY) {
+        return this.t('walletActivity.contractDeployment')
+      }
       if (activity.transaction_category === TOKEN_METHOD_APPROVE) {
         return `${this.t('walletActivity.approved')} ${activity.type_name !== 'n/a' ? activity.type_name.toUpperCase() : activity.type.toUpperCase()}`
       }
@@ -206,7 +213,7 @@ export default {
       return `${`${this.t(activity.action)} ${activity.from}`} `
     },
     getIcon(activity) {
-      if (activity.transaction_category === TOKEN_METHOD_APPROVE) {
+      if ([TOKEN_METHOD_APPROVE, DEPLOY_CONTRACT_ACTION_KEY, CONTRACT_INTERACTION_KEY].includes(activity.transaction_category)) {
         return '$vuetify.icons.coins_approve'
       }
       if (activity.action === ACTIVITY_ACTION_TOPUP) {
