+++ app/src/store/actions.js
@@ -114,7 +114,7 @@ export default {
     resetStore(prefsController.errorStore, errorMessageHandler)
     resetStore(prefsController.paymentTxStore, paymentTxHandler, [])
     resetStore(prefsController.pastTransactionsStore, pastTransactionsHandler, [])
-    resetStore(prefsController.etherscanTxStore, etherscanTxHandler, [])
+    resetStore(txController.etherscanTxStore, etherscanTxHandler, [])
     torus.updateStaticData({ isUnlocked: false })
   },
   setSelectedCurrency({ commit }, payload) {
@@ -373,7 +373,7 @@ export default {
     prefsController.errorStore.subscribe(errorMessageHandler)
     prefsController.paymentTxStore.subscribe(paymentTxHandler)
     prefsController.pastTransactionsStore.subscribe(pastTransactionsHandler)
-    prefsController.etherscanTxStore.subscribe(etherscanTxHandler)
+    txController.etherscanTxStore.subscribe(etherscanTxHandler)
   },
   initTorusKeyring(_, payload) {
     return torusController.initTorusKeyring([payload.privKey], [payload.ethAddress])
